Change the condition for not printing anything in a balance table cell.
Before it was check to see if it's 0.0 and if it was, it left the cell blank
but that is a problem in case the amount owed is less than 0.01 but still
non-zero. So I changed it to check if it's less than 0.01 and if it is, don't
print anything.